Drop cryptographic signatures from converted JAR files
When we change the classes in the JAR files, the cryptographic
signatures will no longer be valid. Deployment of WAR files that
contain those JAR files will fail.

Therefore drop the signatures of any JAR file that is contained
in the WAR.

All dropped signatures and signature files will be logged at leve FINE.
Maybe we should log a warning at the end of the conversion, if signatures
where dropped, to raise more awareness for these kind of modification.