Use commons-compression for the default code path
The JVM code has too much exception throwing and hardcoded behavior
given what the migration tool does, as seen in issue #20. The
commons-compression code does not seem to have the same check in that
case, as seen when using -zipInMemory (it does not do anything special
yet does not complain about duplicate entries since there's no check for
that in the code).
In some cases, this seems faster. In others not so much. This needs more
investigation to verify there's no regression in that area (or anywhere
else).