Copyright - Added an inceptionYear element to the POM file. Updated copyright year in the NOTICE to start with 2020.
Initial commits in this repository and the tag for version 0.0.1 were on January 13th 2020.

Though there was a discussion on the dev@ mailing list that mentioned it earlier
("Proposed plan for Tomcat 10", December 16th 2019)
https://lists.apache.org/thread/8yjg6t383jzdz7k1d1zwzgtvwlt8xjd7

JSTL JARs in Tomcat 10.0 were converted with this tool on January 13th 2020.
https://lists.apache.org/thread/g9zfp511of0hjknm6s9qon40gjdht6w8
https://lists.apache.org/thread/m27zpp3xsz39mzqky21fyxcs6cbr12zy
https://github.com/apache/tomcat/commit/f796b98813d4b12f825276a9a353bf498c64d526

The commit f796b98 above apparently was first authored on December 16th 2019 (according to its "Date" header),
but amended and committed on January 13th 2020.