Updates for javadoc 18+
This is particularly annoying, especially the "warning: use of default
constructor, which does not provide a comment" which means adding the
default constructor just to comment on it (!!). This cannot be disabled
except by filtering out a lot more.
I don't think it is realistic to do this for Tomcat.